Add initial package structure for state, algorithms, and tensor modules. Implement DMRG parameters and workspace structures. Introduce transverse Ising and Heisenberg Hamiltonians. Add environment update functions for MPS. Enhance main module with DMRG comparison tests and SVD functionality.

+++ src/m_tensor/static_tensor.mojo
@@ -8,7 +8,7 @@ from layout.layout import DimList
 from math import ceildiv
 from main import MAX_RANK, list_to_dimlist
 
-alias dtype = DType.float32
+alias dtype = DType.float64
 
 ## General dense tensor structure for block sparse tensor networks
 @fieldwise_init
